requirements.txt
# Core dependencies
pywebview>=4.0
keyboard>=0.13.5


